requirements.txt
azure-functions==1.12.0
WeasyPrint==57.1
